+++ src/ui/button_save.rs
@@ -1,8 +1,8 @@
 use crate::AppState;
 use little_tui::*;
 
-static X: i16 = 75;
-static Y: i16 = 19;
+static X: isize = 75;
+static Y: isize = 19;
 
 pub fn build() -> Element<AppState> {
     let button_save: Element<AppState> = Element::new();
