+++ src/ui/mod.rs
@@ -34,10 +34,10 @@ pub(crate) mod label_end;
 use crate::AppState;
 use little_tui::*;
 
-pub(crate) static APP_WIDTH: u16 = 80; // Width of the game window
-pub(crate) static APP_HEIGHT: u16 = 24; // Height of the game window
+pub(crate) static APP_WIDTH: usize = 80; // Width of the game window
+pub(crate) static APP_HEIGHT: usize = 24; // Height of the game window
 
-pub(crate) fn draw_relative<S>(el: &Element<S>, x: i16, y: i16, state: &AppState) {
+pub(crate) fn draw_relative<S>(el: &Element<S>, x: isize, y: isize, state: &AppState) {
     if Terminal::columns() >= APP_WIDTH && Terminal::rows() >= APP_HEIGHT {
         el.x(state.app_x + x);
         el.y(state.app_y + y);
